:recycle: game切り出し
