+++ kadai3/toshi0607/main.go
@@ -1,51 +1,18 @@
 package main
 
 import (
-	"io"
 	"os"
 	"fmt"
-	"bufio"
-	"time"
+	"github.com/toshi0607/dojo1/kadai3/toshi0607/game"
 )
 
 func main() {
-	ch := input(os.Stdin)
-
-	questions := []string{"a", "b", "c", "d", "e", "f"}
-	timeChan := time.NewTimer(10 * time.Second).C
-
-	for _, v := range questions {
-		fmt.Printf("type %s\n", v)
-		Loop:
-		for {
-			fmt.Print(">")
-			select {
-			case <-timeChan:
-				fmt.Println("Timer expired")
-				return
-			case s := <-ch:
-				if s == v {
-					fmt.Println("ええで")
-					break Loop
-				} else {
-					fmt.Println("あかんで")
-					fmt.Printf("type %s\n", v)
-				}
-			}
-
+	defer func() {
+		if err := recover(); err != nil {
+			fmt.Fprintf(os.Stderr, "Error:\n%s\n", err)
+			os.Exit(1)
 		}
-	}
-	fmt.Println("おしまい")
-}
-
-func input(r io.Reader) <-chan string {
-	ch := make(chan string)
-	go func() {
-		s := bufio.NewScanner(r)
-		for s.Scan() {
-			ch <- s.Text()
-		}
-		close(ch)
 	}()
-	return ch
+	cli := game.Game{OutStream: os.Stdout, ErrStream: os.Stderr}
+	os.Exit(cli.Run())
 }
